removed warning when compiling vector3d with double precision
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@598 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/vector3d.h
@@ -141,7 +141,7 @@ namespace core
 			T l = X*X + Y*Y + Z*Z;
 			if (l == 0)
 				return *this;
-			l = (T) reciprocal_squareroot ( l );
+			l = (T) reciprocal_squareroot ( (f32)l );
 			X *= l;
 			Y *= l;
 			Z *= l;
